[#52118601] Manage /etc/hosts as a file resource
Convert all `/etc/hosts` entries from `host{}` resources to a single
`file{}` resource. This will ensure ensure that our FQDN entry always
appears in the correct position and that `hostname -f` will work as
intended.

As a result, we can no longer support the `purge` param. It is implied (and
documented) that this will remove any existing entries. It won't be possible
to use this in conjunction with any `host{}` resources. For the opinionated
reason that you should be using DNS instead.

Per the previous commit, this also introduces one new entry for
`ip6-allhosts`. This is from Ubuntu stock.